pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "data-hoover"
version = "0.1.0"
description = "Local-first Data Hoover (USGS earthquakes GeoJSON → DuckDB)"
requires-python = ">=3.11"
dependencies = [
  "httpx>=0.27.0",
  "duckdb>=1.0.0",
  "feedparser>=6.0.11",
  "websocket-client>=1.8.0",
  "tomli>=2.0.1; python_version < '3.11'",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
]

[project.scripts]
hoover = "datahoover.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
markers = [
  "network: requires network access",
]
